+++ tests/js/index.js
@@ -49,6 +49,7 @@ var TESTS = {
     MigrationTests: require('./migration-tests'),
     EncryptionTests: require('./encryption-tests'),
     ObjectIDTests: require('./object-id-tests'),
+    AliasTests: require('./alias-tests'),
     // Garbagecollectiontests: require('./garbage-collection'),
 };